+++ src/constants/generals.const.ts
@@ -2,4 +2,5 @@ export const EMPTY_STRING = '';
 export const SPACE_STRING = ' ';
 export const NEWLINE_STRING = '\n';
 export const LOCAL_STORAGE_PREFIX = 'ink-memory:';
+export const DEFAULT_MEMORY_KEY = 'default';
 export const DATA_IMAGE_PNG_PREFIX = 'data:image/png;base64,';